+++ chapter2/pi/v5/main.c
@@ -5,19 +5,19 @@
 
 #define MAX_THREADS		2
 
-int num_steps = 1000000;
+long long num_steps = 1000000;
 double step;
 
 typedef struct {
 	double sum;
-	int start, end;
+	long long start, end;
 } thread_param;
 
 void *thread_func(void *arg)
 {
 	thread_param *thr_arg = (thread_param *) arg;
 	double x, sum = 0.0;
-	int i;
+	long long i;
 
 	for (i = thr_arg->start; i < thr_arg->end; i++) {
 		x = (i + 0.5) * step;
@@ -38,10 +38,10 @@ int main(int argc, char **argv)
 	thread_param thr_arg[MAX_THREADS];
 
 	if (argc > 1)
-		num_steps = atoi(argv[1]);
+		num_steps = atoll(argv[1]);
 	if (num_steps < 100)
 		num_steps = 1000000;
-	printf("\nnum_steps = %d\n", (int)num_steps);
+	printf("\nnum_steps = %lld\n", num_steps);
 
 	gettimeofday(&start, NULL);
